+++ 10_benchmark_perf.py
@@ -4,23 +4,22 @@ Benchmark script using hydra.
 Run it with python benchmark.py --config-name ismrm2024 to reproduce results of abstract.
 
 """
-import warnings
 import csv
-import hydra
 import logging
-import time
 import os
-import numpy as np
-from itertools import product
-from omegaconf import DictConfig
+import time
+import warnings
 from pathlib import Path
 
-from mrinufft import get_operator
-from mrinufft.trajectories.io import read_trajectory
-from hydra_callbacks.monitor import ResourceMonitorService
+import hydra
+import numpy as np
 from hydra_callbacks.logger import PerfLogger
+from hydra_callbacks.monitor import ResourceMonitorService
+from mrinufft import get_operator
+from mrinufft.io import read_trajectory
+from omegaconf import DictConfig
 
-from utils import get_smaps
+AnyShape = tuple[int, int, int]
 
 
 CUPY_AVAILABLE = True
@@ -81,7 +80,7 @@ def get_data(cfg):
 
 
 @hydra.main(
-    config_path=".",
+    config_path="perf",
     config_name="benchmark_config",
     version_base=None,
 )
@@ -141,7 +140,7 @@ def main_app(cfg: DictConfig) -> None:
                 elif task == "adjoint":
                     nufft.adj_op(ksp_data)
                 elif task == "grad":
-                    nufft.get_grad(data, ksp_data)
+                    nufft.data_consistency(data, ksp_data)
                 else:
                     raise ValueError(f"Unknown task {task}")
             toc = time.perf_counter()
@@ -172,6 +171,87 @@ def main_app(cfg: DictConfig) -> None:
     if CUPY_AVAILABLE:
         cp.get_default_memory_pool().free_all_blocks()
 
+
+def get_smaps(
+    shape: AnyShape,
+    n_coils: int,
+    antenna: str = "birdcage",
+    dtype: np.dtype = np.complex64,
+    cachedir="/tmp/smaps",
+) -> np.ndarray:
+    """Get sensitivity maps for a specific antenna.
+
+    Parameters
+    ----------
+    shape
+        Volume shape
+    n_coils
+        number of coil in the antenna
+    antenna
+        name of the antenna to emulate. Only "birdcage" is currently supported.
+    dtype
+        return datatype for the sensitivity maps.
+    """
+    if antenna == "birdcage":
+        try:
+            os.makedirs(cachedir, exist_ok=True)
+            smaps = np.load(f"{cachedir}/smaps_{n_coils}_{shape}.npy")
+        except FileNotFoundError:
+            smaps = _birdcage_maps((n_coils, *shape), nzz=n_coils, dtype=dtype)
+            np.save(f"{cachedir}/smaps_{n_coils}_{shape}.npy", smaps)
+        return smaps
+    else:
+        raise NotImplementedError
+
+
+def _birdcage_maps(
+    shape: AnyShape, r: float = 1.5, nzz: int = 8, dtype: np.dtype = np.complex64
+) -> np.ndarray:
+    """Simulate birdcage coil sensitivies.
+
+    Parameters
+    ----------
+    shape
+        sensitivity maps shape (nc, x,y,z)
+    r
+        Relative radius of birdcage.
+    nzz
+        number of coils per ring.
+    dtype
+
+    Returns
+    -------
+    np.ndarray: complex sensitivity profiles.
+
+    References
+    ----------
+    https://sigpy.readthedocs.io/en/latest/_modules/sigpy/mri/sim.html
+    """
+    if len(shape) == 4:
+        nc, nz, ny, nx = shape
+    elif len(shape) == 3:
+        nc, ny, nx = shape
+        nz = 1
+    else:
+        raise ValueError("shape must be [nc, nx, ny, nz] or [nc, nx, ny]")
+    c, z, y, x = np.mgrid[:nc, :nz, :ny, :nx]
+
+    coilx = r * np.cos(c * (2 * np.pi / nzz), dtype=np.float32)
+    coily = r * np.sin(c * (2 * np.pi / nzz), dtype=np.float32)
+    coilz = np.floor(np.float32(c / nzz)) - 0.5 * (np.ceil(nc / nzz) - 1)
+    coil_phs = np.float32(-(c + np.floor(c / nzz)) * (2 * np.pi / nzz))
+
+    x_co = (x - nx / 2.0) / (nx / 2.0) - coilx
+    y_co = (y - ny / 2.0) / (ny / 2.0) - coily
+    z_co = (z - nz / 2.0) / (nz / 2.0) - coilz
+    rr = (x_co**2 + y_co**2 + z_co**2) ** 0.5
+    phi = np.arctan2(x_co, -y_co) + coil_phs
+    out = (1 / rr) * np.exp(1j * phi)
+
+    rss = sum(abs(out) ** 2, 0) ** 0.5
+    out /= rss
+    out = np.squeeze(out)
+    return out.astype(dtype)
 
 if __name__ == "__main__":
     main_app()
