+++ solver_utils.py
